# docs/partners/openapi/chatbot-v1.json
{
  "components": {
    "schemas": {
      "ChatMessage": {
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "user or assistant",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "chat_history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prior turns for this request (canonical).",
            "title": "Chat History"
          },
          "conversation_history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deprecated alias for chat_history.",
            "title": "Conversation History"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "User message (alias for query)",
            "title": "Message"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "User message (alias for message)",
            "title": "Query"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional product user id for Langfuse analytics (client-supplied until auth).",
            "title": "User Id"
          },
          "user_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfile"
              },
              {
                "type": "null"
              }
            ],
            "description": "User profile: country, plan_type, category."
          }
        },
        "title": "ChatRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SessionCreateRequest": {
        "properties": {
          "category": {
            "enum": [
              "Government",
              "NGOs",
              "Agribusinesses",
              "Farmers"
            ],
            "title": "Category",
            "type": "string"
          }
        },
        "required": [
          "category"
        ],
        "title": "SessionCreateRequest",
        "type": "object"
      },
      "SessionCreateResponse": {
        "properties": {
          "category": {
            "enum": [
              "Government",
              "NGOs",
              "Agribusinesses",
              "Farmers"
            ],
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "created_at",
          "category"
        ],
        "title": "SessionCreateResponse",
        "type": "object"
      },
      "SessionStatusResponse": {
        "description": "Response for GET /v1/sessions/{session_id}.",
        "properties": {
          "alive": {
            "title": "Alive",
            "type": "boolean"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "has_summary": {
            "default": false,
            "title": "Has Summary",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "turn_count": {
            "default": 0,
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "alive"
        ],
        "title": "SessionStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Public v1 API for Ask ADZA enterprise partners. Authenticate with X-API-Key or Authorization: Bearer when enterprise auth is enabled.",
    "title": "Ask ADZA Enterprise Chatbot API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat": {
      "post": {
        "description": "POST /v1/chat \u2014 Generic chat endpoint (backward compatible).\n\nPrefer the plan-scoped routes (/v1/chat/{plan}) for new integrations.\nplan_type is read from user_profile.plan_type when using this endpoint.",
        "operationId": "v1_chat_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat"
      }
    },
    "/v1/chat/agribusinesses": {
      "post": {
        "description": "POST /v1/chat/agribusinesses \u2014 Agribusinesses tier (cross-country, market volatility).",
        "operationId": "v1_chat_agribusinesses_v1_chat_agribusinesses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Agribusinesses"
      }
    },
    "/v1/chat/farmers": {
      "post": {
        "description": "POST /v1/chat/farmers \u2014 Farmers tier (localized crop/rainfall/market).",
        "operationId": "v1_chat_farmers_v1_chat_farmers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Farmers"
      }
    },
    "/v1/chat/free": {
      "post": {
        "description": "POST /v1/chat/free \u2014 Free tier (single country, top-line answers only).",
        "operationId": "v1_chat_free_v1_chat_free_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Free"
      }
    },
    "/v1/chat/government": {
      "post": {
        "description": "POST /v1/chat/government \u2014 Government tier (national/sub-national, food security).",
        "operationId": "v1_chat_government_v1_chat_government_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Government"
      }
    },
    "/v1/chat/integrated": {
      "post": {
        "description": "POST /v1/chat/integrated \u2014 Integrated tier (full access, category lens per message).",
        "operationId": "v1_chat_integrated_v1_chat_integrated_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Integrated"
      }
    },
    "/v1/chat/ngos": {
      "post": {
        "description": "POST /v1/chat/ngos \u2014 NGOs tier (multi-region risk, program angles).",
        "operationId": "v1_chat_ngos_v1_chat_ngos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Chat Ngos"
      }
    },
...
    "/v1/sessions": {
      "post": {
        "operationId": "v1_create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Create Session"
      }
    },
    "/v1/sessions/{plan_type_slug}": {
      "post": {
        "description": "Create a session scoped to a specific plan type.\n\nThe plan_type is inferred from the URL slug (e.g. /v1/sessions/farmers).\nCategory defaults to the natural category for the plan; can be overridden\nby passing ?category=... as a query parameter in future.",
        "operationId": "v1_create_plan_session_v1_sessions__plan_type_slug__post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_type_slug",
            "required": true,
            "schema": {
              "title": "Plan Type Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Create Plan Session"
      }
    },
    "/v1/sessions/{session_id}": {
      "delete": {
        "description": "DELETE /v1/sessions/{session_id} \u2014 explicitly clear a session (logout / clear chat).\n\nIdempotent: returns success even if the session was already gone.",
        "operationId": "v1_delete_session_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Delete Session"
      },
      "get": {
        "description": "GET /v1/sessions/{session_id} \u2014 check if a session is alive.\n\nReturns session metadata (category, turn count, whether a summary exists).\nReturns 404 if the session has expired or never existed.",
        "operationId": "v1_get_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Get Session"
      }
    },
...
  }
}